+++ config.py
@@ -1,6 +1,10 @@
 # This configuration file is for Gunicorn in a Linux production environment. For Windows systems, please disregard it#
 # import multiprocessing
 # workers = multiprocessing.cpu_count() * 2 + 1
+import os
+
+os.environ["GUNICORN_RUNNING"] = "1"
+
 bind = "127.0.0.1:5000"
 
 workers = 1