+++ myapp/admin.py
@@ -15,7 +15,6 @@ from myapp.db_model import (
     Guarantee,
     Option,
     Question,
-    QuestionCategory,
     QuestionImgURL,
     Response,
     ResponseDetail,
@@ -33,9 +32,11 @@ from myapp.mail import send_mail, survey_result_mail
 from myapp.survey_utils import (
     DCQuestion,
     build_dc_questions,
+    build_dict_question,
     get_response_total_score,
     is_survey_mounted,
     is_survey_response_expired,
+    user_answered_question,
 )
 from myapp.utils import (
     build_pagination_dict,
@@ -81,7 +82,7 @@ def set_config():
     key = str(data.get("key") or "").strip()
     value = str(data.get("value") or "").strip()
     type_ = str(data.get("type") or "").strip()
-    description = str(data.get("description") or "").strip()
+    description = str(data.get("desc") or "").strip()
 
     if not key:
         return jsonify({"code": 1, "desc": "need key!"})
@@ -158,7 +159,7 @@ def mod_survey():
     data = request.get_json()
     id_ = int(data.get("id") or 0)
     name = str(data.get("name") or "")
-    desc = str(data.get("desc") or "")
+    desc = str(data.get("description") or "")
 
     if id_ == 0 or not name:
         return jsonify({"code": 1, "desc": "必须填写名称!"})
@@ -182,7 +183,11 @@ def add_question():
     前端提供一个题目列表，每个列表元素包含：问卷ID、题目标题、类型、分数、选项和排序 ID 六个参数
     排序 ID 为 0 代表题目加入到末尾，其他值则为插入
     """
-    data = request.get_json()
+    data: dict[str, Any] | None = request.get_json(silent=True)
+
+    if data is None:
+        return jsonify({"code": 4, "desc": "数据格式有误"}), 400
+
     question_list = data.get("questions", [])
     survey_id = data.get("surveyId", None)
 
@@ -202,6 +207,9 @@ def add_question():
                 target_display_order=question.display_order,
             )
 
+            db.session.add(new_question)
+            db.session.flush()
+
             option_objs = [
                 Option(question_id=new_question.id, option_text=opt.text, is_correct=opt.is_correct)
                 for opt in question.options
@@ -212,7 +220,6 @@ def add_question():
                 for img in question.images
             ]
 
-            db.session.add(new_question)
             db.session.add_all(option_objs)
             db.session.add_all(image_objs)
 
@@ -276,11 +283,12 @@ def edit_question():
     """
     data = request.get_json()
     question_data = data.get("question", None)
+    survey_id = question_data.get("surveyId", None)
 
     if question_data is None:
         return jsonify({"code": 1, "desc": "fail"})
 
-    is_valid, error_info_or_list = build_dc_questions([question_data], question_data.get("surveyId", None))
+    is_valid, error_info_or_list = build_dc_questions([question_data], survey_id)
     if not is_valid:
         return jsonify({"code": 1, "desc": error_info_or_list})
 
@@ -683,52 +691,98 @@ def get_responses():
 @required_role("admin")
 def get_survey(sid: int):
     """
-    查询指定问卷
+    管理员预览问卷
     """
     survey: Survey | None = db.session.get(Survey, sid)
     if not survey:
-        return jsonify({"code": 1, "desc": "未找到问卷"}), 404
+        return jsonify({"code": 1, "desc": "问卷不存在"})
 
     survey_data: dict[str, Any] = {
         "id": survey.id,
         "name": survey.name,
         "description": survey.description,
-        "create_time": parse_dt_to_iso_utc(survey.create_time),
+        "createTime": parse_dt_to_iso_utc(survey.create_time),
+        "questions": [
+            build_dict_question("admin-view", question)
+            for question in survey.questions
+            if not question.logical_deletion
+        ],
+    }
+
+    return jsonify({"code": 0, "desc": "success", "data": survey_data})
+
+
+@admin.route("/detail/<int:resp_id>", methods=["GET"])
+@login_required
+@required_role("admin")
+def review_survey(resp_id: int):
+    """
+    批改问卷
+    """
+    res: Response | None = db.session.get(Response, resp_id)
+    if res is None:
+        return jsonify({"code": 1, "desc": "答卷不存在"})
+
+    survey: Survey | None = db.session.get(Survey, res.survey_id)
+
+    if survey is None:
+        return jsonify({"code": 1, "desc": "未找到答卷对应的问卷"})
+
+    survey_data: dict[str, Any] = {
+        "id": res.id,
+        "name": survey.name,
+        "description": survey.description,
+        "createTime": parse_dt_to_iso_utc(survey.create_time),
+        "isReviewed": res.is_reviewed,
         "questions": [],
     }
 
     # 查询问卷中的所有题目
     for question in survey.questions:
-        # 不返回被逻辑删除的题目
-        if question.logical_deletion:
+        response_score: ResponseScore | None = db.session.execute(
+            select(ResponseScore).where(ResponseScore.question_id == question.id, ResponseScore.response_id == resp_id)
+        ).scalar()
+
+        # 查询题目中的所有选项详情，多选题会创建多个ResponseDetail，每个里面的answer是一个选择的选项ID
+        stmt = select(ResponseDetail).where(
+            ResponseDetail.question_id == question.id, ResponseDetail.response_id == resp_id
+        )
+        details = db.session.execute(stmt).scalars().all()
+
+        # 如果题目被逻辑删除，并且用户未作答，则不显示
+        if question.logical_deletion and not user_answered_question(details):
             continue
 
-        question_data: dict[str, Any] = {
-            "display_order": question.display_order,
-            "id": question.id,
-            "title": question.question_text,
-            "type": question.question_type,
-            "score": question.score,
-            "images": [],
-            "options": [],
-        }
+        survey_data["questions"].append(build_dict_question("admin-review", question, details, response_score))
+    return jsonify({"code": 1, "desc": "未找到答卷对应的问卷", "data": survey_data})
 
-        for img in question.img_list:
-            question_data["img_list"].append({"id": img.id, "alt": img.img_alt, "data": img.img_data})
 
-        # 查询题目中的所有选项
-        for option in question.options:
-            question_data["options"].append(
-                {
-                    "id": option.id,
-                    "text": option.option_text,
-                    "isCorrect": option.is_correct,
-                }
-            )
+@admin.route("/detail_score", methods=["POST"])
+@login_required
+@required_role("admin")
+def set_score():
+    """
+    批改某个题目，前端提供题目ID，答卷ID，和分数
+    """
+    req_data = request.json
+    if req_data:
+        score = req_data.get("score")
+        question_id = req_data.get("questionId")
+        response_id = req_data.get("responseId")
+        if not all([score, response_id, question_id]):
+            return jsonify({"code": 2, "desc": "字段无效！"}), 400
 
-        survey_data["questions"].append(question_data)
-
-    return jsonify({"code": 0, "desc": "success", "data": survey_data})
+        stmt = select(ResponseScore).where(
+            ResponseScore.question_id == question_id, ResponseScore.response_id == response_id
+        )
+        res = db.session.scalar(stmt)
+        if res is not None:
+            res.score = score
+        else:
+            db.session.add(ResponseScore(score, question_id, response_id))
+        db.session.commit()
+        return jsonify({"code": 0, "desc": "批改成功！"})
+    return jsonify({"code": 1, "desc": "缺少信息！"})
 
 
 @admin.route("/reviewed", methods=["POST"])
@@ -744,7 +798,7 @@ def reviewed_response():
         raw_status: int | None = req_data.get("status", None)
 
         if rid is None or raw_status is None:
-            return jsonify({"code": 1, "desc": "缺少参数! "})
+            return jsonify({"code": 1, "desc": "缺少参数! "}), 400
 
         try:
             status: ResponseStatus = ResponseStatus(raw_status)
@@ -793,119 +847,6 @@ def reviewed_response():
         return jsonify({"code": 0, "desc": "操作成功"})
 
     return jsonify({"code": 4, "desc": "缺少数据! "})
-
-
-@admin.route("/detail/<int:resp_id>", methods=["GET"])
-@login_required
-@required_role("admin")
-def get_detail(resp_id: int):
-    """
-    查询指定问卷
-    """
-    res: Response | None = db.session.get(Response, resp_id)
-    if res is None:
-        return jsonify({"code": 1, "desc": "答卷不存在"})
-
-    survey: Survey | None = db.session.get(Survey, res.survey_id)
-
-    if survey is None:
-        return jsonify({"code": 1, "desc": "未找到问卷"})
-
-    survey_data: dict[str, Any] = {
-        "id": res.id,
-        "name": survey.name,
-        "description": survey.description,
-        "create_time": parse_dt_to_iso_utc(survey.create_time),
-        "isReviewed": res.is_reviewed,
-        "questions": [],
-    }
-
-    # 查询问卷中的所有题目
-    for question in survey.questions:
-        response_score: ResponseScore | None = db.session.execute(
-            select(ResponseScore).where(ResponseScore.question_id == question.id, ResponseScore.response_id == resp_id)
-        ).scalar()
-
-        # 查询题目中的所有选项详情
-        stmt = select(ResponseDetail).where(
-            ResponseDetail.question_id == question.id, ResponseDetail.response_id == resp_id
-        )
-        details = db.session.execute(stmt).scalars().all()
-
-        # 如果题目被逻辑删除，并且用户未作答，则不显示
-        if question.logical_deletion and len(details) == 0:
-            continue
-
-        score = 0
-        user_selected_option: list[int] = []
-
-        if response_score is not None:
-            score = response_score.score
-
-        question_data: dict[str, Any] = {
-            "display_order": question.display_order,
-            "id": question.id,
-            "title": question.question_text,
-            "type": question.question_type,
-            "score": question.score,
-            "userGetScore": score,
-            "options": [],
-            "img_list": [],
-            "text_answer": "",
-        }
-
-        for img in question.img_list:
-            question_data["img_list"].append({"alt": img.img_alt, "data": img.img_data})
-
-        # 标注用户选择的选项
-        if (
-            question.question_type == QuestionCategory.SINGLE_CHOICE.value
-            or question.question_type == QuestionCategory.MULTIPLE_CHOICE.value
-        ):
-            for detail in details:
-                user_selected_option.append(int(detail.answer))
-
-        for option in question.options:
-            question_data["options"].append(
-                {
-                    "id": option.id,
-                    "text": option.option_text,
-                    "isCorrect": option.is_correct,
-                    "isSelected": True if option.id in user_selected_option else False,
-                    "inputText": details[0].answer if question.question_type in [3, 4] and len(details) != 0 else "",
-                }
-            )
-
-        survey_data["questions"].append(question_data)
-    return jsonify(survey_data)
-
-
-@admin.route("/detail_score", methods=["POST"])
-@login_required
-@required_role("admin")
-def set_score():
-    """
-    批改某个题目，前端提供题目ID，答卷ID，和分数
-    """
-    req_data = request.json
-    if req_data:
-        score = req_data.get("score")
-        question_id = req_data.get("questionId")
-        response_id = req_data.get("responseId")
-        if not all([score, response_id, question_id]):
-            return jsonify({"code": 2, "desc": "字段无效！"}), 400
-
-        stmt = select(ResponseScore).where(
-            ResponseScore.question_id == question_id, ResponseScore.response_id == response_id
-        )
-        res = db.session.scalar(stmt)
-        if res is not None:
-            res.score = score
-        else:
-            db.session.add(ResponseScore(score, question_id, response_id))
-        db.session.commit()
-        return jsonify({"code": 0, "desc": "批改成功！"})
-    return jsonify({"code": 1, "desc": "缺少信息！"})
 
 
 @admin.route("/slot/add", methods=["POST"])